seccion de jugadores completados

+++ app/src/main/java/com/luisro00005513/crossoversports/Entities/FavoritoXUsuario.java
@@ -15,9 +15,9 @@ public class FavoritoXUsuario {
     private  String userName;
 
     @ColumnInfo(name = "player_id")
-    private Integer playerId;
+    private Long playerId;
 
-    public FavoritoXUsuario(String userName, Integer playerId) {
+    public FavoritoXUsuario(String userName, Long playerId) {
         this.userName = userName;
         this.playerId = playerId;
     }
@@ -38,11 +38,11 @@ public class FavoritoXUsuario {
         this.userName = userName;
     }
 
-    public Integer getPlayerId() {
+    public Long getPlayerId() {
         return playerId;
     }
 
-    public void setPlayerId(Integer playerId) {
+    public void setPlayerId(Long playerId) {
         this.playerId = playerId;
     }
 }
